+++ EisenhowerMatrix/Enums/QuartersName.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace EisenhowerMain
+namespace EisenhowerMain.Enums
 {
     public enum QuartersName
     {
