+++ EisenhowerMatrix/Input.cs
@@ -13,51 +13,31 @@ namespace EisenhowerMain
 
         public string InputTitle()
         {
-            string title = Console.ReadLine();
-            return title;
+            return Console.ReadLine();
         }
 
         public string InputDeadline()
         {
             string pattern = @"^(0[1-9]|[12][0-9]|3[01])-(0[1-9]|1[0-2])$";
             string deadline = Console.ReadLine();
-            if (Regex.IsMatch(deadline, pattern))
-            {
-                return deadline;
-            }
-            return "dd-MM";
-
+            return Regex.IsMatch(deadline, pattern) ? deadline : "dd-MM";
         }
 
         public string InputImportance()
         {
             string pattern = @"^[YN]$";
             string importance = Console.ReadLine().ToUpper();
-            if (Regex.IsMatch(importance, pattern))
-            {
-                return importance;
-            }
-            return "Y/N";
+            return Regex.IsMatch(importance, pattern) ? importance : "Y/N";
         }
 
         public bool InputConfirmation() 
         {
             ConsoleKeyInfo confirmationKey = Console.ReadKey(intercept: true);
-
-            if (confirmationKey.Key == ConsoleKey.Y)
-            {
-                Console.Clear();
-                Console.WriteLine("Task deleted.");
-                Thread.Sleep(1000);
-                return true;
-            }
-            else
-            {
-                Console.Clear();
-                Console.WriteLine("Deletion cancelled.");
-                Thread.Sleep(1000);
-                return false;
-            }
+            bool confirmation = confirmationKey.Key == ConsoleKey.Y;
+            Console.Clear();
+            Console.WriteLine(confirmation ? "Task deleted." : "Deletion cancelled.");
+            Thread.Sleep(1000);         
+            return confirmation;
         }
